pyproject.toml
[project]
name = "3dbag-docs"
version = "2024.12.16"
description = "3DBAG documentation"
readme = "README.md"
license = "CC-BY 4.0"
license-files = ["LICEN[CS]E*"]
authors = [
    { name = "Balázs Dukai", email = "balazs.dukai@3dgi.nl" },
    { name = "Ravi Peters", email = "ravi.peters@3dgi.nl" },
    { name = "Gina Stavropoulou", email = "G.Stavropoulou@tudelft.nl" },
]
requires-python = ">=3.11"
dependencies = [
    "bag3d-specs",
    "mkdocs-material~=9.6.15",
    "tabulate>=0.9.0",
]

[tool.uv.sources]
bag3d-specs = { git = "https://github.com/3DBAG/3dbag-specs.git", branch = "develop" }
